+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.97.12", date: "2026-03-24" };
+  return { version: "8.97.13", date: "2026-03-24" };
 }